fix: public-cloud-terraform: remove provider/alias informations from examples
alias and provider keyword are an expert way of doing Terraform, and is
complexifying the quick start example for no added value.

Customers that add resource must take this complexity into account and
can lead to complicated errors.
Removing the complexity is better, as advanced users that want to deal
with multiple configuration of the same provider know how to do that.

Signed-off-by: Romain Beuque <556072+rbeuque74@users.noreply.github.com>