ESXi 8 reintroduction + updates
Guids detailing the installation and configuration of ESXi have been updated to better reflect ESXi 8, as it is being reintroduced starting on the 15th of September 2025.

The "End of Support" guide has thus been removed, and disclaimers about the introduction also have been added to the existing guides.

The ESXi partitioning guide has been renamed to reflect that it both refers to the installation of ESXi and the partitioning necessary for that installation.